+++ ArchipelagoSurvivors/Patches/SurvivorScreenPatch.cs
@@ -4,8 +4,10 @@ using Il2CppVampireSurvivors.UI;
 using UnityEngine;
 using UnityEngine.UI;
 using static ArchipelagoSurvivors.APSurvivorClient;
+using static ArchipelagoSurvivors.InformationTransformer;
 using static ArchipelagoSurvivors.Patches.PlayerPatch;
 using Il2CppGeneric = Il2CppSystem.Collections.Generic;
+using Object = UnityEngine.Object;
 
 namespace ArchipelagoSurvivors.Patches;
 
@@ -19,6 +21,9 @@ public static class SurvivorScreenPatch
     public static TickBoxController HurryController;
 
     public static TickBoxController ArcaneController;
+
+    private static GameObject StageSelectButton;
+    private static int LastBeaten = -1;
 
     [HarmonyPatch(typeof(CharacterSelectionPage), "Start"), HarmonyPostfix]
     public static void OverrideCharacterStart(CharacterSelectionPage __instance)
@@ -39,19 +44,20 @@ public static class SurvivorScreenPatch
         {
             LastMinuteCheck = -1;
         }
-        
+
         EggController.Update();
 
         foreach (var (character, ui) in __instance._characterItemUIs)
         {
             ui.gameObject.SetActive(AllowedCharacters.Contains(character));
-            ui.GetChild(3).GetComponent<Image>().enabled = !CharactersBeaten.Contains(character);
+            ui._LockIcon.enabled = !IsHurryLocked && !CharactersBeaten.Contains(character);
         }
     }
 
     [HarmonyPatch(typeof(StageSelectPage), "Start"), HarmonyPostfix]
     public static void OverrideMapStart(StageSelectPage __instance)
     {
+        LastBeaten = -1;
         var infoBox = __instance.GetChild(0).GetChild(2).GetChild(0).GetChildren();
         HyperController = infoBox[2];
         HyperController.VariableTracker = "hyper";
@@ -59,6 +65,7 @@ public static class SurvivorScreenPatch
         HurryController.VariableTracker = "hurry";
         ArcaneController = infoBox[4];
         ArcaneController.VariableTracker = "arcanas";
+        StageSelectButton = __instance.GetChild(2);
     }
 
     [HarmonyPatch(typeof(StageSelectPage), "Update"), HarmonyPostfix]
@@ -67,11 +74,60 @@ public static class SurvivorScreenPatch
         HyperController.Update();
         HurryController.Update();
         ArcaneController.Update();
-        
-        foreach (var stage in __instance.GetChild(0).GetChild(1).GetChild(0).GetChild(0).GetComponentsInChildren<StageItemUI>())
+
+        var missingText = Client!.MissingLocations.Select(loc => Client!.DataLookup.Locations[loc]).ToArray();
+        foreach (var stage in __instance.GetChild(0)
+                                        .GetChild(1)
+                                        .GetChild(0)
+                                        .GetChild(0)
+                                        .GetComponentsInChildren<StageItemUI>())
         {
+            var hasBeaten = !StagesBeaten.Contains(stage.Type);
+            var stageName = StageTypeToName[stage.Type];
+
             stage.gameObject.SetActive(AllowedStages.Contains(stage.Type));
-            stage.GetChild(5).SetActive(!StagesBeaten.Contains(stage.Type));
+            stage._Exclamation.SetActive(!IsHurryLocked && hasBeaten);
+            
+            if (stage.Type == StageType.MACHINE)
+            {
+                if (APSurvivorClient.GoalRequirement == GoalRequirement.KillTheDirector)
+                {
+                    stage.DescriptionText.text = $"Stages beaten requirement to open:\n[{StagesBeaten.Count}] of [{StagesToBeatForDirector}]";
+                }
+            }
+            else
+            {
+                var chestChecksMissing = missingText.Where(s => !s.Contains("Beaten") && s.Contains(stageName))
+                                                    .Select(s => s.Replace($" on {stageName}", ""))
+                                                    .ToArray();
+                
+                var enemyChecksMissing = missingText.Where(s => s.Contains("Kill"))
+                                                    .Select(s => EnemyNameToType[s[5..]])
+                                                    .Where(et => EnemyStages[et].Contains(stage.Type))
+                                                    .Select(et => EnemyTypeToName[et])
+                                                    .ToArray();
+                
+                if (chestChecksMissing.Any())
+                {
+                    stage.DescriptionText.text = $"Chest Checks Remaining:\n{string.Join(", ", chestChecksMissing)}";
+                }
+                else if (enemyChecksMissing.Any())
+                {
+                    stage.DescriptionText.text = $"Enemy Checks Remaining:\n{string.Join(", ", enemyChecksMissing)}";
+                }
+                else
+                {
+                    stage.DescriptionText.text = "All Chest and Enemy Checks Got!";
+                }
+            }
+            
+            if (stage.Type != StageType.MACHINE ||
+                APSurvivorClient.GoalRequirement != GoalRequirement.KillTheDirector) continue;
+
+            if (__instance._selectedStage == stage)
+            {
+                StageSelectButton.SetActive(StagesBeaten.Count >= StagesToBeatForDirector);
+            }
         }
     }
 }
@@ -89,7 +145,7 @@ public class TickBoxController
             "eggs" => !IsEggesLocked,
             _ => false
         };
-    
+
     private TickBoxUI Box = null;
     private GameObject Gobj = null;